Refactor import paths in BillPage for improved organization and clarity
